requirements.txt
APScheduler>=3.0.5
Babel>=2.1.1
CairoSVG>=1.0.19
DecoratorTools>=1.8
Jinja2>=2.8
Landscape-Client>=14.12
Mako>=1.0.3
MarkupSafe>=0.23
PAM>=0.4.2
Paste>=2.0.2
PasteDeploy>=1.5.2
Pillow>=3.0.0
PyMySQL>=0.6.7
PyYAML>=3.11
Routes>=2.2
SOAPpy>=0.12.0
SQLAlchemy>=1.0.9
Tempita>=0.5.2
Twisted-Core>=13.2.0
WSME>=0.8.0
WebOb>=1.5.1
WebTest>=2.0.20
aioeventlet>=0.4
alembic>=0.8.3
amqp>=1.4.8
anyjson>=0.3.3
apt-xapian-index>=0.45
argparse>=1.2.1
beautifulsoup4>=4.4.1
cachetools>=1.1.5
cairocffi>=0.7.2
cffi>=1.3.1
chardet>=2.0.1
cliff>=1.4.5
cmd2>=0.6.7
colorama>=0.2.5
configobj>=4.7.2
contextlib2>=0.4.0
cssselect>=0.9.1
ctypeslib>=0.5.6
deap>=0.7.1
debiancontributors>=0.6
debtagshw>=0.1
debtcollector>=1.1.0
decorator>=4.0.6
defer>=1.0.6
defusedxml>=0.4.1
demjson>=1.6
dexml>=0.5.1
enum34>=1.1.1
eventlet>=0.17.4
fasteners>=0.14.1
fpconst>=0.7.2
futures>=3.0.3
futurist>=0.8.0
greenlet>=0.4.9
html5lib>=0.999
influxdb>=2.10.0
iso8601>=0.1.11
jsonschema>=2.3.0
keystoneauth1>=2.1.0
keystonemiddleware>=4.0.0
kombu>=3.0.30
logutils>=0.3.3
lxml>=3.5.0
mock>=1.0.1
monotonic>=0.4
msgpack-python>=0.4.6
netaddr>=0.7.18
netifaces>=0.10.4
ordereddict>=1.1
oslo.concurrency>=3.1.0
oslo.config>=3.1.0
oslo.context>=1.0.0
oslo.db>=4.1.0
oslo.i18n>=3.1.0
oslo.log>=2.1.0
oslo.messaging>=3.1.0
oslo.middleware>=3.1.0
oslo.serialization>=2.1.0
oslo.service>=1.1.0
oslo.utils>=3.2.0
pbr>=1.8.1
pecan>=1.0.3
ply>=3.4
prettytable>=0.7.2
pyOpenSSL>=0.13
pycadf>=2.0.1
pycparser>=2.10
pycups>=1.9.66
pycurl>=7.19.3
pygal>=2.0.11
pyinotify>=0.9.6
pyliblzma>=0.5.3
pyparsing>=2.0.1
pyserial>=2.6
pysqlite>=1.0.1
python-apt>=0.9.3.5ubuntu1
python-dateutil>=2.4.2
python-debian>=0.1.21-nmu2ubuntu2
python-debianbts>=1.11
python-designateclient>=1.0.0
python-editor>=0.5
python-keystoneclient>=2.0.0
python-memcached>=1.57
python-novaclient>=2.35.0
pytz>=2015.7
report>=0.0.1.dev45
reportlab>=3.2.0
repoze.lru>=0.6
requests>=2.8.1
retrying>=1.3.3
simplegeneric>=0.8.1
simplejson>=3.8.1
singledispatch>=3.4.0.3
six>=1.10.0
sqlalchemy-migrate>=0.10.0
sqlparse>=0.1.18
ssh-import-id>=3.21
stevedore>=1.10.0
tinycss>=0.3
trollius>=2.0
tzlocal>=1.2
urlgrabber>=3.9.1
urllib3>=1.7.1
waitress>=0.8.10
wheel>=0.24.0
wrapt>=1.10.6
wsgiref>=0.1.2
yum-metadata-parser>=1.1.4
zope.interface>=4.0.5
